Incorporo mejoras en estructura y distribucion de contenido en archivos, Torres Cintia Cristina
