+++ src/assets/components/App.jsx
@@ -1,5 +1,4 @@
 import { Routes, Route, Link } from 'react-router-dom';
-import { useState, useEffect, useCallback } from 'react';
 
 import Navbar from 'react-bootstrap/Navbar';
 import Nav from 'react-bootstrap/Nav';
@@ -8,95 +7,20 @@ import Container from "react-bootstrap/esm/Container";
 import AlumnoFormulario from './AlumnoFormulario';
 import AlumnoDelete from './AlumnoDelete';
 import EditarAlumno from './EditarAlumno';
-import Home from './Home';
-import Nosotros from './Nosotros';
-
-import MostrarListaAlumnos from './MostrarListaAlumnos';
+import Home from '../pages/Home';
+import Nosotros from '../pages/Nosotros';
+import MostrarListaAlumnos from '../pages/MostrarListaAlumnos';
 import BuscarAlumnoPage from './BuscarAlumnoPage';
-import BuscarAlumno from './BuscarAlumno';
 
-import AlumnoData from './AlumnoData';
+import useAlumnos from '../hooks/useAlumnos';
 
 function App() {
-
-  const [alumnos, setAlumnos] = useState(() => [
-    AlumnoData({
-      nombre: "Ana",
-      apellido: "López",
-      curso: "3°A",
-      email: "ana.lopez@hotmail.com",
-      domicilio: "Calle Falsa 123",
-      telefono: "4230957",
-      carrera: "APU",
-      estado: true,
-    }),
-    AlumnoData({
-      nombre: "Carlos",
-      apellido: "Pérez",
-      curso: "2°B",
-      email: "carlos.perez@gmail.com",
-      domicilio: "Av. Siempre Viva 742",
-      telefono: "388-4842277",
-      carrera: "APU",
-      estado: true,
-    }),
-    AlumnoData({
-      nombre: "Lucía",
-      apellido: "Martínez",
-      curso: "1°C",
-      email: "lucia.martinez@gmail.com",
-      domicilio: "San Martín 500",
-      telefono: "388-4348958",
-      carrera: "APU",
-      estado: true,
-    }),
-    AlumnoData({
-      nombre: "Javier",
-      apellido: "Gómez",
-      curso: "4°D",
-      email: "javier.gomez@example.com",
-      domicilio: "Belgrano 1200",
-      telefono: "388-5253053",
-      carrera: "APU",
-      estado: true,
-    }),
-  ]);
-
-  useEffect(() => {
-    if (alumnos.length > 0) {
-      console.log("Lista de alumnos actualizada:", alumnos);
-    }
-  }, [alumnos]);
-
-  const handleAgregarAlumno = useCallback((nuevoAlumno) => {
-    setAlumnos(prev => [...prev, nuevoAlumno]);
-  }, []);
-
-  const deshabilitarAlumno = useCallback((valorBusqueda) => {
-    const alumnoEncontrado = BuscarAlumno(alumnos, valorBusqueda);
-
-    if (alumnoEncontrado && alumnoEncontrado.estado !== false) {
-      setAlumnos(prev =>
-        prev.map(alumno =>
-          alumno === alumnoEncontrado ? { ...alumno, estado: false } : alumno
-        )
-      );
-      return true;
-    }
-    return false;
-  }, [alumnos]);
-
-
-  const actualizarAlumno = useCallback((alumnoActualizado) => {
-    setAlumnos(prev =>
-      prev.map(alumno =>
-        alumno.lu === alumnoActualizado.lu
-          ? { ...alumno, ...alumnoActualizado }
-          : alumno
-      )
-    );
-  }, []);
-
+  const {
+    alumnos,
+    handleAgregarAlumno,
+    deshabilitarAlumno,
+    actualizarAlumno
+  } = useAlumnos();
 
   return (
     <div>
