+++ src/assets/pages/DetalleAlumno.jsx
